+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Go server template: add `go_server_pubsub` outputter to support direct
+  Pub/Sub delivery ([DENG-9533](https://mozilla-hub.atlassian.net/browse/DENG-9533))
+
 ## 19.1.1
 
 - Use JSON serialization for on-disk caching